✨ :: LazyColumn 으로 변경

+++ app/src/main/java/com/g3c1/oasis_android/feature_seat/data/data_source_impl/SeatDataSourceImpl.kt
@@ -39,7 +39,7 @@ class SeatDataSourceImpl @Inject constructor(
         }.flowOn(Dispatchers.IO)
     }
 
-    override suspend fun patchSeatData(seatId: Int): Flow<ApiState<Unit>> {
+    override suspend fun patchSeatData(seatId: Long): Flow<ApiState<Unit>> {
         return flow {
             try {
                 val response = service.patchSeatData(seatId)
