+++ app/src/main/java/com/g3c1/oasis_android/feature_seat/domain/repository/SeatRepository.kt
@@ -7,5 +7,5 @@ import kotlinx.coroutines.flow.Flow
 interface SeatRepository {
     suspend fun getSeatDataRepository(): Flow<ApiState<List<SeatDTO>>>
 
-    suspend fun patchSeatDataRepository(seatId: Int): Flow<ApiState<Unit>>
+    suspend fun patchSeatDataRepository(seatId: Long): Flow<ApiState<Unit>>
 }
